Command pattern added
